Add files via upload
This code is a script for predicting the close prices of a stock for the next 30 days using the historical data of the stock for the last year. The script prompts the user to enter the stock ticker symbol, then it uses the yfinance library to download the historical data for the stock from the last year. It then processes the data by converting the date column to a datetime object, setting it as the index, sorting it by date, and calculating the 200-day moving average. The script uses the HistGradientBoostingRegressor from the scikit-learn library to fit a model to the data, and make predictions for the next 30 days. It then adds some random noise to the predictions and plots them along with the actual close prices for the last year. It also plots 200 day moving average. Finally, it sets the x-axis label, y-axis label, plot title, and displays the plot.